+++ src/backend/core/tests/test_services_ai_services.py
@@ -2,7 +2,6 @@
 Test ai API endpoints in the impress core app.
 """
 
-import json
 from unittest.mock import MagicMock, patch
 
 from django.core.exceptions import ImproperlyConfigured
@@ -58,9 +57,8 @@ def test_api_ai__client_error(mock_create):
 def test_api_ai__client_invalid_response(mock_create):
     """Fail when the client response is invalid"""
 
-    answer = {"no_answer": "This is an invalid response"}
     mock_create.return_value = MagicMock(
-        choices=[MagicMock(message=MagicMock(content=json.dumps(answer)))]
+        choices=[MagicMock(message=MagicMock(content=None))]
     )
 
     with pytest.raises(
@@ -77,49 +75,10 @@ def test_api_ai__client_invalid_response(mock_create):
 def test_api_ai__success(mock_create):
     """The AI request should work as expect when called with valid arguments."""
 
-    answer = '{"answer": "Salut"}'
     mock_create.return_value = MagicMock(
-        choices=[MagicMock(message=MagicMock(content=answer))]
+        choices=[MagicMock(message=MagicMock(content="Salut"))]
     )
 
     response = AIService().transform("hello", "prompt")
 
     assert response == {"answer": "Salut"}
-
-
-@override_settings(
-    AI_BASE_URL="http://example.com", AI_API_KEY="test-key", AI_MODEL="test-model"
-)
-@patch("openai.resources.chat.completions.Completions.create")
-def test_api_ai__success_sanitize(mock_create):
-    """The AI response should be sanitized"""
-
-    answer = '{"answer": "Salut\\n \tle \nmonde"}'
-    mock_create.return_value = MagicMock(
-        choices=[MagicMock(message=MagicMock(content=answer))]
-    )
-
-    response = AIService().transform("hello", "prompt")
-
-    assert response == {"answer": "Salut\n \tle \nmonde"}
-
-
-@override_settings(
-    AI_BASE_URL="http://example.com", AI_API_KEY="test-key", AI_MODEL="test-model"
-)
-@patch("openai.resources.chat.completions.Completions.create")
-def test_api_ai__success_when_sanitize_fails(mock_create):
-    """The AI request should work as expected even with badly formatted response."""
-
-    # pylint: disable=C0303
-    answer = """{ 
-        "answer"        :       
-        "Salut le monde"
-    }"""
-    mock_create.return_value = MagicMock(
-        choices=[MagicMock(message=MagicMock(content=answer))]
-    )
-
-    response = AIService().transform("hello", "prompt")
-
-    assert response == {"answer": "Salut le monde"}